Split :readmore and :readmore-material modules again
