+++ readmore-material/README.md
@@ -1,0 +1,48 @@
+# ReadMoreText
+
+<a href="https://jitpack.io/#fornewid/metronome"><img src="https://jitpack.io/v/fornewid/metronome.svg"/></a>
+
+## Installation
+
+```gradle
+repositories {
+    maven { url "https://jitpack.io" }
+}
+
+dependencies {
+    implementation "com.github.fornewid.metronome:readmore-material:<version>"
+}
+```
+
+Depend on the [latest version](https://github.com/fornewid/metronome/releases) of the library.
+
+## Usage
+
+#### [ReadMoreText](https://github.com/fornewid/metronome/blob/main/readmore-material/src/main/java/soup/metronome/readmore/ReadMoreText.kt)
+
+`ReadMoreText` is based on material theme like `androidx.compose.material.Text`.
+
+```kotlin
+@Composable
+fun ReadMoreTextSample() {
+    val (expanded, onExpandedChange) = rememberSaveable { mutableStateOf(false) }
+    ReadMoreText(
+        text = "...",
+        expanded = expanded,
+        modifier = Modifier
+            .clickable { onExpandedChange(!expanded) }
+            .fillMaxWidth(),
+        readMoreText = "Read more",
+        readMoreColor = Color.Black,
+        readMoreFontSize = 15.sp,
+        readMoreFontStyle = FontStyle.Normal,
+        readMoreFontWeight = FontWeight.Bold,
+        readMoreFontFamily = FontFamily.Default,
+        readMoreTextDecoration = TextDecoration.None,
+        readMoreMaxLines = 2,
+        readMoreStyle = SpanStyle(
+            // ...
+        )
+    )
+}
+```
